+++ pytc_gui/TO-DO
@@ -1,1 +1,2 @@
 truncate white space from end of ##csv spew
+relink AIC test to put spew in text box
