+++ pytc_gui/widgets/experiment_box/__init__.py
@@ -5,4 +5,4 @@ Submodule holding the experimental box visual elements of the gui.
 __author__ = "Michael J. Harms"
 __date__ = "2017-06-01"
 
-from .exp_box import ExperimentBox
+from .experiment_box import ExperimentBox
