+++ pytc_gui/widgets/experiment_box/experiment_box.py
